Moved CommandDebugPopup and LocalizationDebugWindow into new Debug folder and kept LocalizationDebugWindow ViewModel
