+++ src/PixiEditor/Views/Dialogs/DebugDialogs/CommandDebugPopup.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="PixiEditor.Views.Dialogs.CommandDebugPopup"
+﻿<Window x:Class="PixiEditor.Views.Dialogs.DebugDialogs.CommandDebugPopup"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
